refactor: RoomRepository 내 Querydsl 이 아닌 JPQL 로 쿼리 작성
