+++ backend/src/main/java/com/codesquad/airbnb/room/RoomService.java
@@ -23,7 +23,9 @@ public class RoomService {
     }
 
     public RoomDetailResponse findRoom(Integer roomId) {
-        Room room = roomRepository.findById(roomId);
+        Room room = roomRepository.findByIdWithDetail(roomId)
+            .orElseThrow(() -> new IllegalStateException("숙소 정보가 존재하지 않습니다."));
+
         return RoomDetailResponse.from(room);
     }
